Several minor bug fixes
* Makefile (check): Added kernel/pkg/headers/create_link to the files
  tested when doing a make check.

* kernel/pkg/image/prerm: Use debconf to ask questions.
  Removed a bunch of obsolete variables.
  One of the questions now asked using debconf is the one when we are
  trying to remove the kernel version we are running, the other is the
  question asked if we have a valid entry for the version being removed
  in the boot loader configuration.

* kernel/ruleset/minimal.mk (minimal_debian): Do not ever write the
  control or the changelog if they are already there. (probably the
  official version).

* kernel/ruleset/targets/headers.mk (install/$(h_package)): create the
  /etc/kernel/postinst.d directory, and install the script create
  link-VERSION there.
  (debian/$(h_package)): Add the /etc/kernel/postinst.d/create_link-*
  file to conffiles.

git-archimport-id: srivasta@debian.org--etch/kernel-package--devel--9.0--patch-58

+++ kernel/ruleset/minimal.mk
@@ -82,18 +82,20 @@ debian: minimal_debian
 minimal_debian:
 	$(REASON)
 	test -d debian || mkdir debian
-	sed -e 's/=V/$(version)/g'         -e 's/=D/$(debian)/g'        \
-	       -e 's/=A/$(DEB_HOST_ARCH)/g'   -e 's/=SA/$(INT_SUBARCH)/g'  \
-                -e 's/=L/$(int_loaderdep) /g' -e 's/=I/$(initrddep)/g'     \
-                -e 's/=CV/$(VERSION).$(PATCHLEVEL)/g'                      \
-                -e 's/=M/$(maintainer) <$(email)>/g'                       \
-                -e 's/=ST/$(INT_STEM)/g'      -e 's/=B/$(KERNEL_ARCH)/g'   \
+	test -f debian/control || sed         -e 's/=V/$(version)/g'        \
+                -e 's/=D/$(debian)/g'         -e 's/=A/$(DEB_HOST_ARCH)/g'  \
+	        -e 's/=SA/$(INT_SUBARCH)/g'   -e 's/=L/$(int_loaderdep) /g' \
+                -e 's/=I/$(initrddep)/g'                                    \
+                -e 's/=CV/$(VERSION).$(PATCHLEVEL)/g'                       \
+                -e 's/=M/$(maintainer) <$(email)>/g'                        \
+                -e 's/=ST/$(INT_STEM)/g'      -e 's/=B/$(KERNEL_ARCH)/g'    \
 		         $(CONTROL) > debian/control
-	sed -e 's/=V/$(version)/g' -e 's/=D/$(debian)/g'                   \
-	    -e 's/=A/$(DEB_HOST_ARCH)/g' -e 's/=M/$(maintainer) <$(email)>/g' \
-            -e 's/=ST/$(INT_STEM)/g'     -e 's/=B/$(KERNEL_ARCH)/g'           \
-		$(LIBLOC)/changelog > debian/changelog
-	install -p -m 755 $(LIBLOC)/rules debian/rules
+	test -f debian/changelog ||  sed -e 's/=V/$(version)/g'             \
+	    -e 's/=D/$(debian)/g'        -e 's/=A/$(DEB_HOST_ARCH)/g'       \
+            -e 's/=ST/$(INT_STEM)/g'     -e 's/=B/$(KERNEL_ARCH)/g'         \
+	    -e 's/=M/$(maintainer) <$(email)>/g' 	                    \
+             $(LIBLOC)/changelog > debian/changelog
+	test -f debian/rules || install -p -m 755 $(LIBLOC)/rules debian/rules
 	for file in $(DEBIAN_FILES); do                                      \
             cp -f  $(LIBLOC)/$$file ./debian/;                               \
         done
